+++ src/distribution/tests/ParetoTest.cpp
@@ -63,7 +63,7 @@ ParetoTest::testIt()
     Pareto dis(config.get("dist"));
 
     Average<double> average;
-    for(int32_t ii = 0; ii < 1000000; ++ii) {
+    for(long int ii = 0; ii < 1000000; ++ii) {
         average.put(dis());
     }
 
@@ -83,7 +83,7 @@ ParetoTest::testVar()
     Pareto dis = Pareto(3.0, 10.0);
 
     VarEstimator var;
-    for(int32_t ii = 0; ii < 1000000; ++ii) {
+    for(long int ii = 0; ii < 1000000; ++ii) {
         var.put(dis());
     }
