+++ src/distribution/tests/WeibullTest.cpp
@@ -59,7 +59,7 @@ WeibullTest::testIt()
   Weibull dis(config.get("dist"));
 
   Average<double> average;
-  for(int32_t i = 0; i < 1000000; ++i)
+  for(long int i = 0; i < 1000000; ++i)
     {
       average.put((dis)());
     }
@@ -75,7 +75,7 @@ WeibullTest::testVar()
   Weibull dis = Weibull(2.0, 1.0);
 
   VarEstimator var;
-  for(int32_t i = 0; i < 1000000; ++i)
+  for(long int i = 0; i < 1000000; ++i)
     {
       var.put(dis());
     }
